REF: RangeIndex tests (#33050)
